+++ app/src/main/java/com/github/se/assocify/ui/screens/treasury/TreasuryGraph.kt
@@ -6,7 +6,7 @@ import com.github.se.assocify.navigation.Destination
 import com.github.se.assocify.navigation.NavigationActions
 import com.github.se.assocify.ui.screens.treasury.accounting.balance.balanceDetailedGraph
 import com.github.se.assocify.ui.screens.treasury.accounting.budget.budgetDetailedGraph
-import com.github.se.assocify.ui.screens.treasury.receipt.receiptGraph
+import com.github.se.assocify.ui.screens.treasury.receiptstab.receipt.receiptGraph
 
 fun NavGraphBuilder.treasuryGraph(navigationActions: NavigationActions) {
   composable(
